Change image manipulation to use complex numbers
Change the image manipulation module to use complex numbers instead of
tuple[int, int] for representing coordinates. For this, add two utility
functions: to_complex and to_int_tuple. They do exactly what they say
they do. That is, convert between tuple[int, int] and complex.

The motivation behind this change is that complex numbers are easier
to scale and add to each other than tuples. If operations were
vectorized on tuples in Python, then this change would be unnecessary.
However, they are not. One possible alternative would have been to use
numpy, for which operations are vectorized, but that would be more
complicated than this solution.