Add API Convenient For Programs
Add an API that is convenient for programs to use. This is a generator
that also gives a list of the images used in each card. The purpose of
this is to allow applications to revise decks after-the-fact, if the
user does not like the current card.

This version does not use file I/O, but rather takes parameters and is a
generator. It is a generator so that applications can display progress
bars other than the text-based one provided by the regular main()
function, which is now renamed to deck() to avoid confusion, but goes
otherwise unchanged.